find_options: replace with module version
