+++ lib/entries.js
@@ -2,7 +2,7 @@
 
 var es = require('event-stream');
 var sgvdata = require('sgvdata');
-var traverse = require('traverse');
+var find_options = require('./query');
 var ObjectID = require('mongodb').ObjectID;
 
 /**********\
@@ -10,61 +10,11 @@ var ObjectID = require('mongodb').ObjectID;
  * Encapsulate persistent storage of sgv entries.
 \**********/
 
-function build_query (opts) {
-
-  opts =  walker({ date: parseInt, sgv: parseInt })(opts);
-  return opts;
-}
-
-function walker (spec) {
-  var fns = [ ];
-  for (var prop in spec) {
-    var typer = spec[prop];
-    fns.push(walk_prop(prop, typer));
-  }
-
-  function exec (obj) {
-    var fn;
-    while (fn = fns.shift( )) {
-      obj = fn(obj);
-    }
-    return obj;
-  }
-  return exec
-}
-
-function walk_prop (prop, typer) {
-  function iter (opts) {
-    if (opts && opts.find && opts.find[prop]) {
-      traverse(opts.find[prop]).forEach(function (x) {
-        if (this.isLeaf) {
-          this.update(typer(x));
-        }
-      });
-    }
-    return opts;
-  }
-  return iter;
-}
-
-var TWO_DAYS = 172800000;
 function storage(env, ctx) {
 
   // TODO: Code is a little redundant.
 
   var with_collection = ctx.store.with_collection(env.mongo_collection);
-
-  function find_options (opts) {
-    var finder = build_query(opts);
-    var query = finder && finder.find ? finder.find : { };
-    if (!query.date && !query.dateString) {
-      query.date = { $gte: Date.now( ) - ( TWO_DAYS * 2 ) };
-    }
-    if (query._id && query._id.length) {
-      query._id = ObjectID(query._id);
-    }
-    return query;
-  }
 
   // query for entries from storage
   function list (opts, fn) {
